Исправление опечатки в пространстве имен/Correcting a typo in the namespace.
