+++ src/DataServer.Core/DataServer.Core.csproj
@@ -21,6 +21,7 @@
   </ItemGroup>
 
   <ItemGroup>
+    <Folder Include="Logging\" />
     <Folder Include="Properties\" />
   </ItemGroup>
