+++ src/DataServer.Core/Locking/LockerService.cs
@@ -1,0 +1,49 @@
+﻿/*
+* Copyright 2024 SpTSoftware
+* Licensed under the Apache License, Version 2.0 (the "License");
+* you may not use this file except in compliance with the License.
+* You may obtain a copy of the License at
+* 
+*     http://www.apache.org/licenses/LICENSE-2.0
+* 
+* Unless required by applicable law or agreed to in writing, software
+* distributed under the License is distributed on an "AS IS" BASIS,
+* WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+* See the License for the specific language governing permissions and
+* limitations under the License.
+*/
+
+using DataServer.Core.Notifications;
+
+namespace DataServer.Core.Locking
+{
+	public class LockerService : ILockerService
+	{
+		private readonly INotificationsService _NotificationsService;
+
+		public LockerService(INotificationsService notificationsService)
+		{
+			this._NotificationsService = notificationsService;
+		}
+
+		public void Run()
+		{
+			this._NotificationsService.Run();
+		}
+
+		public void Run(params object[] @params)
+		{
+			this._NotificationsService.Run(@params);
+		}
+
+		public void Stop()
+		{
+			this._NotificationsService.Stop();
+		}
+
+		public void Stop(params object[] @params)
+		{
+			this._NotificationsService.Stop(@params);
+		}
+	}
+}
